third_party/blink/web_tests/platform/generic/accessibility/aria-hidden-with-elements-expected.txt

cell	cell	cell
cell	cell	cell
cell
test
test
option
option
heading

This tests that aria-hidden works as expected on elements that are subclasses of AccessibilityRenderObject

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS child_1.childrenCount is 1
PASS grandchild_1.isIgnored is true
PASS grandchild_1.childrenCount is 2
PASS grandchild_1.childAtIndex(0).isIgnored is true
PASS grandchild_1.childAtIndex(1).isIgnored is true
PASS child_2.childAtIndex(0).isIgnored is true
PASS child_3.childrenCount is 2
PASS child_3.childAtIndex(0).isIgnored is true
PASS child_3.childAtIndex(1).isIgnored is true
PASS child_4.childrenCount is 1
PASS child_4.role is 'AXRole: AXGenericContainer'
PASS child_4.isIgnored is false
PASS child_4.childAtIndex(0).isIgnored is true
PASS child_5.isIgnored is false
PASS child_5.role is 'AXRole: AXHeading'
PASS successfullyParsed is true

TEST COMPLETE
